# perf: preallocate buffers in nodejs.