--- a/random.js
+++ b/random.js
@@ -1,1 +1,8 @@
-module.exports = require('crypto').randomBytes
+var crypto = require('crypto')
+
+module.exports = function (bytes) {
+  var buffer = Buffer.allocUnsafe(bytes)
+  crypto.randomFillSync(buffer)
+
+  return buffer
+}