Improve perfomance for nodejs version. My results:

#### before
```
nanoid                331,738 ops/sec
nanoid/generate       327,488 ops/sec
uid.sync              327,499 ops/sec
uuid/v4               321,873 ops/sec
shortid                31,894 ops/sec

Async:
uid                    61,713 ops/sec
nanoid/async           60,515 ops/sec

Non-secure:
rndm                2,310,600 ops/sec
nanoid/non-secure   2,441,124 ops/sec
```

#### after
```
nanoid                374,446 ops/sec
nanoid/generate       367,777 ops/sec
uid.sync              326,813 ops/sec
uuid/v4               322,756 ops/sec
shortid                31,409 ops/sec

Async:
uid                    62,463 ops/sec
nanoid/async           66,546 ops/sec

Non-secure:
rndm                2,323,620 ops/sec
nanoid/non-secure   2,463,917 ops/sec
```