@ai 2018-09-26
Wow, it's amazing.

I will accept it soon and fix test coverage