@ai 2018-09-26
Released in 1.2.4. Thanks :+1: